Default route LSA refresh
Fix LSA refresh for redistributed LSAs. After default route
configuration and setup as distributed, it is distributed as expected.
But after LSA refresh timer expires it is not refreshed and the
corresponding LSA is remove from LSADB. This problem relates to all
LSAs distributed as AS-External.

Tags: fix, dev, usr

Change-Id: I6d04d2c5da1cbf55707b53eed7a53042c06d55ba
Signed-off-by: Nikolay Stroykov <stroykov@mera.ru>